Rename jtwStrategy module directory to jwt
