+++ packages/api/src/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 
 import { GithubModule } from './github/github.module';
-import { JWTModule } from './jwtStrategy/jwt.module';
+import { JWTModule } from './jwt/jwt.module';
 
 @Module({
   imports: [PassportModule.register({ defaultStrategy: 'jwt' }), JWTModule, GithubModule],
