+++ packages/api/src/auth/github/github.controller.spec.ts
@@ -4,8 +4,8 @@ import { Test } from '@nestjs/testing';
 import type { User } from '../../../../shared/src/models/user';
 import { PrismaService } from '../../prisma/prisma.service';
 import { UsersRepository } from '../../users/users.repository';
-import { JWTModule } from '../jwtStrategy/jwt.module';
-import { JwtStrategy } from '../jwtStrategy/jwt.strategy';
+import { JWTModule } from '../jwt/jwt.module';
+import { JwtStrategy } from '../jwt/jwt.strategy';
 import { GithubStrategy } from './github.strategy';
 import { GithubController } from './github.controller';
 import { GithubUserData } from './github.model';
