+++ packages/api/src/auth/github/github.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 
 import { UsersModule } from '../../users/users.module';
 import { UsersRepository } from '../../users/users.repository';
-import { JWTModule } from '../jwtStrategy/jwt.module';
-import { JwtStrategy } from '../jwtStrategy/jwt.strategy';
+import { JWTModule } from '../jwt/jwt.module';
+import { JwtStrategy } from '../jwt/jwt.strategy';
 import { GithubController } from './github.controller';
 import { GithubStrategy } from './github.strategy';
